+++ components/utils.tsx
@@ -7,7 +7,7 @@ export function NotLoggedIn() {
             <div className="medSep" />
             <div className="text-5xl font-bold">Oh snap... You're not logged in!</div>
             <div className="smallSep" />
-            <p className="font-semibold">It seems that you're not logged into our services! You must be logged in to access dashboard features.</p>
+            <p className="font-semibold text-light">It seems that you're not logged into our services! You must be logged in to access dashboard features.</p>
             <a href="/api/oauth" className="block bg-white text-black font-bold transition duration-200 shadow hover:shadow-2xl ease-in-and-out rounded-lg mx-auto mt-16 px-2 py-4 text-xl font-sans w-60">
                 Log In
             </a>
@@ -24,7 +24,7 @@ function ErrorBase(props: {
         return (
             <>
                 <div className="text-5xl font-bold">Oh snap... It crashed!</div>
-                <p className="font-semibold">Try it again later!</p>
+                <p className="font-semibold text-light">Try it again later!</p>
             </>
         );
     } else {
@@ -74,13 +74,14 @@ export function Error(props?: {
 }
 
 interface SVGProps {
-    size?: "small" | "medium" | "large" | "semiLarge" | "semiMedium";
+    size?: "small" | "medium" | "large" | "semiLarge" | "semiMedium" | "reallyLarge" | "superLarge" | "guildIcon";
     color?: HexColor | string;
+    className?: string;
 }
 
 export function ExternalIcon(props?: SVGProps) {
     return (
-        <svg className={`icon outbound svg-${props.size}`} xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false" x="0px" y="0px" viewBox="0 0 100 100" width="15" height="15">
+        <svg className={`icon outbound svg-${props.size} ${props.className == null ? "" : props.className}`} xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false" x="0px" y="0px" viewBox="0 0 100 100" width="15" height="15">
             <path fill={props.color || "currentColor"} d="M18.8,85.1h56l0,0c2.2,0,4-1.8,4-4v-32h-8v28h-48v-48h28v-8h-32l0,0c-2.2,0-4,1.8-4,4v56C14.8,83.3,16.6,85.1,18.8,85.1z"></path>
             <polygon fill={props.color || "currentColor"} points="45.7,48.7 51.3,54.3 77.2,28.5 77.2,37.2 85.2,37.2 85.2,14.9 62.8,14.9 62.8,22.9 71.5,22.9"></polygon>
         </svg>
@@ -105,5 +106,15 @@ export function ExternalLink({ children, href, blurple, className }: ExternalLin
             {children}
             <ExternalIcon />
         </a>
+    );
+}
+
+interface SomeSVGProps {
+    color?: HexColor | string;
+}
+
+export function UnknownIcon(props: SVGProps){
+    return (
+        <svg className={`svg-${props.size} ${props.className == null ? "" : props.className}`} fill={props.color || "none"} viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path d="M12 2c5.523 0 10 4.478 10 10s-4.477 10-10 10S2 17.522 2 12 6.477 2 12 2Zm0 13.5a1 1 0 1 0 0 2 1 1 0 0 0 0-2Zm0-8.75A2.75 2.75 0 0 0 9.25 9.5a.75.75 0 0 0 1.493.102l.007-.102a1.25 1.25 0 1 1 2.5 0c0 .539-.135.805-.645 1.332l-.135.138c-.878.878-1.22 1.447-1.22 2.53a.75.75 0 0 0 1.5 0c0-.539.135-.805.645-1.332l.135-.138c.878-.878 1.22-1.447 1.22-2.53A2.75 2.75 0 0 0 12 6.75Z" fill="#5865f2"/></svg>
     );
 }